+++ src/Copse.CodeGen/AsyncToSync.cs
@@ -300,7 +300,6 @@ namespace Copse.CodeGen
         ["Copse.Linq.Async"] = "Copse.Linq.Treenumerators",
         ["Copse.Linq.Async.Treenumerators"] = "Copse.Linq.Treenumerators",
         ["Copse.Linq.Async.Treenumerables"] = "Copse.Linq.Treenumerables",
-        ["Copse.Async.Topologies"] = "Copse.Topologies",
         ["Copse.Async.ChildEnumerators"] = "Copse.ChildEnumerators",
         ["Copse.Linq.Async.Stores"] = "Copse.Linq.Stores",
       };