+++ src/Copse.Linq.Async/Treenumerables/Walkable/AsyncPruneDescendantsWhereWalkable.cs
@@ -1,6 +1,5 @@
 using Copse;
 using Copse.Core;
-using Copse.Topologies;
 using System;
 using System.Threading.Tasks;