+++ src/Copse.Linq.Tests/TreeWalkerLawTests.cs
@@ -1,5 +1,4 @@
 using Copse.Core;
-using Copse.Topologies;
 using Copse.Linq.Treenumerables;
 using Copse;
 using Copse.SimpleSerializer;