+++ src/Copse.Linq/TreenumerableExtensions/WalkableTreenumerable/Treenumerable.Extend.g.cs
@@ -4,7 +4,6 @@
 // </auto-generated>
 using Copse.Core;
 using Copse;
-using Copse.Topologies;
 using Copse.Linq.Treenumerables;
 using System;