+++ src/Copse.Linq/Treenumerables/Walkable/LazyTopology.g.cs
@@ -4,7 +4,7 @@
 // </auto-generated>
 using Copse.Core;
 
-namespace Copse.Topologies
+namespace Copse.Linq.Treenumerables
 {
   // TreeTopology.Lazy's engine: a walkable's topology, call-by-need. Anything that must
   // build lazily over "whatever topology this walkable's door will bind" -- a view whose